+++ sudokuplus.koplugin/_meta.lua
@@ -9,5 +9,5 @@ return {
     name = "sudokuplus",
     fullname = _("Sudoku+"),
     description = _([[A logical Sudoku puzzle game and tutor for e-ink readers.]]),
-    version = "1.0.0",
+    version = "1.1.0",
 }